docs: added usage with nuxt fix: exported LokiOptions fix: improved error logging when a fetch failed
