+++ src/index.ts
@@ -1,1 +1,8 @@
 export { LokiReporter } from "./reporters/loki";
+
+export interface LokiOptions {
+  baseURL: string;
+  user?: string;
+  token?: string;
+  labels?: Record<string, string>;
+}
