--- src/types.ts
@@ -1,6 +1,0 @@
-export interface LokiOptions {
-  baseURL: string;
-  user?: string;
-  token?: string;
-  labels?: Record<string, string>;
-}
